feature: add an option to not require checksum
By default sentences without a checksum are considered to have an
invalid checksum (like before). With the new option `requireChecksum`
set to false, they will be accepted but we will still validate checksums
on lines that have one.

Tests for this are in signalk-parser-nmea0183 (this is where existing
tests were already placed).